+++ src/utils/compact.ts
